chore(postgrest): move update-json to scripts
